src/backend/requirements.txt
setuptools==68.0.0
Werkzeug==2.3.6
gunicorn==20.1.0
Jinja2==3.1.2
Flask==2.3.2
python-dotenv==1.0.0
psycopg2==2.9.6
sqlalchemy==2.0.19
sqlalchemy-utils==0.41.1